requirements.txt
# Core dependencies
python-telegram-bot==20.3
supabase==2.0.0
httpx==0.24.1
aiohttp==3.9.0
APScheduler==3.10.4
python-dotenv==1.0.0
nest_asyncio==1.5.8

# Voice recognition dependencies
openai>=1.0.0
groq>=0.4.0
aiofiles>=23.0.0

# Type hints and validation
typing-extensions==4.8.0

# Monitoring and logging
sentry-sdk[logging]==1.45.0
structlog==23.2.0

# Development dependencies
ruff==0.1.9
mypy==1.8.0
bandit==1.7.5

# Testing dependencies  
pytest==7.4.4
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-cov==4.1.0
coverage==7.3.4
